requirements.txt
PyYAML
elasticsearch_async
elasticsearch==6.3.1